Move ( cluster ) repository implementation into the adapter package
