+++ cmd/pipeline/main.go
@@ -926,7 +926,7 @@ func main() {
 
 					integratedServiceOperationDispatcher := integratedserviceadapter.MakeCadenceIntegratedServiceOperationDispatcher(workflowClient, commonLogger)
 					integratedServiceManagerRegistry := integratedservices.MakeIntegratedServiceManagerRegistry(integratedServiceManagers)
-					clusterRepository := integratedservices.NewClusterRepository(clusterManager.KubeConfigFunc(), externaldns.NewSpecWrapper())
+					clusterRepository := integratedserviceadapter.NewClusterRepository(clusterManager.KubeConfigFunc(), externaldns.NewSpecWrapper())
 					integratedServicesService = integratedservices.NewISServiceV2(integratedServiceManagerRegistry, integratedServiceOperationDispatcher, clusterRepository, commonLogger)
 				} else {
 					// legacy setup
