+++ cmd/worker/main.go
@@ -587,7 +587,7 @@ func main() {
 			var featureRepository integratedservices.IntegratedServiceRepository
 			featureRepository = integratedserviceadapter.NewGormIntegratedServiceRepository(db, logger)
 			if config.IntegratedService.V2 {
-				featureRepository = integratedservices.NewClusterRepository(clusterManager.KubeConfigFunc(), externaldns.NewSpecWrapper())
+				featureRepository = integratedserviceadapter.NewClusterRepository(clusterManager.KubeConfigFunc(), externaldns.NewSpecWrapper())
 			}
 
 			kubernetesService := kubernetes.NewService(
